+++ app/scripts/shell/dimSearchFilter.directive.js
@@ -53,7 +53,15 @@
       stattype: ['intellect', 'discipline', 'strength'],
       inloadout: ['inloadout'],
       new: ['new'],
-      glimmer: ['glimmeritem', 'glimmerboost', 'glimmersupply']
+      glimmer: ['glimmeritem', 'glimmerboost', 'glimmersupply'],
+      hasLight: ['light', 'haslight'],
+      weapon: ['weapon'],
+      armor: ['armor'],
+      cosmetic: ['cosmetic'],
+      equipment: ['equipment', 'equippable'],
+      postmaster: ['postmaster', 'inpostmaster'],
+      equipped: ['equipped'],
+      transferable: ['transferable', 'movable']
     };
 
     var keywords = _.flatten(_.flatten(_.values(filterTrans)).map(function(word) {
@@ -537,6 +545,47 @@
       },
       new: function(predicate, item) {
         return item.isNew;
+      },
+      hasLight: function(predicate, item) {
+        const lightBuckets = ["BUCKET_CHEST",
+                                 "BUCKET_LEGS",
+                                 "BUCKET_ARTIFACT",
+                                 "BUCKET_HEAVY_WEAPON",
+                                 "BUCKET_PRIMARY_WEAPON",
+                                 "BUCKET_CLASS_ITEMS",
+                                 "BUCKET_SPECIAL_WEAPON",
+                                 "BUCKET_HEAD",
+                                 "BUCKET_ARMS",
+                                 "BUCKET_GHOST"];
+        return item.bucket && _.contains(lightBuckets, item.bucket.id);
+      },
+      weapon: function(predicate, item) {
+        return item.bucket && item.bucket.sort === 'Weapons';
+      },
+      armor: function(predicate, item) {
+        return item.bucket && item.bucket.sort === 'Armor';
+      },
+      cosmetic: function(predicate, item) {
+        const cosmeticBuckets = ["BUCKET_SHADER",
+                                 "BUCKET_MODS",
+                                 "BUCKET_EMOTES",
+                                 "BUCKET_EMBLEM",
+                                 "BUCKET_VEHICLE",
+                                 "BUCKET_SHIP",
+                                 "BUCKET_HORN"];
+        return item.bucket && _.contains(cosmeticBuckets, item.bucket.id);
+      },
+      equipment: function(predicate, item) {
+        return item.equipment;
+      },
+      postmaster: function(predicate, item) {
+        return item.location && item.location.inPostmaster;
+      },
+      equipped: function(predicate, item) {
+        return item.equipped;
+      },
+      transferable: function(predicate, item) {
+        return !item.notransfer;
       }
     };
   }